Live commenting on posts
Using action cable implemented a minimal version of live
commenting on the posts. The post id is used as room to
identify the post on which comments are added or deleted.
Multiple people who are viewing the post can now see the
comments posted by other users on the same post in real time.
The feature is available only when the user is logged in.

Isssue #13