Display bookmarks differenly than other posts
Bookmarks were created in the same way as that of posts with a minor difference, ie; the bookmark link is stored
on the content field of the post. To navigate to the bookmark properly the display logic is updated to link and
open the bookmark on a separate tab.

Issue #16